GH-20 Fix rendering of Lobby and Game
