+++ src/game/components/Board.js
@@ -1,5 +1,7 @@
 import React from 'react';
-import { GridList, GridTile } from 'material-ui/GridList';
+import { GridList, GridListTile, GridListTileBar } from 'material-ui/GridList';
+import IconButton from 'material-ui/IconButton';
+import InfoIcon from 'material-ui-icons/Info';
 
 const styles = {
   root: {
@@ -24,18 +26,25 @@ export default class Board extends React.PureComponent {
       <div style={styles.root}>
         { words
           ? <GridList
-            cols={5}
-            padding={2}
-            cellHeight="auto"
-            style={styles.gridList}
-          >
-            {words.map(word =>
-              <GridTile
-                key={word.position}
-                title={word.word}
-              />
-            )}
-          </GridList>
+              cols={5}
+              spacing={2}
+              cellHeight="auto"
+              style={styles.gridList}
+            >
+              {words.map(word =>
+                <GridListTile key={word.position}>
+                  <GridListTileBar
+                    title={word.word}
+                    subtitle={<span>TODO: votes</span>}
+                    actionIcon={
+                      <IconButton>
+                        <InfoIcon />
+                      </IconButton>
+                    }
+                  />
+                </GridListTile>
+              )}
+            </GridList>
           : null
         }
       </div>
