+++ src/lobby/actionTypes.js
@@ -1,4 +1,8 @@
-export const GET_PLAYER_LIST = 'GET_PLAYER_LIST';
-export const UPDATE_PLAYER = 'UPDATE_PLAYER';
-export const UPDATE_GAME = 'UPDATE_GAME';
-export const START_GAME = 'START_GAME';
+const actionTypes = {
+  GET_PLAYER_LIST: 'GET_PLAYER_LIST',
+  UPDATE_PLAYER: 'UPDATE_PLAYER',
+  UPDATE_GAME: 'UPDATE_GAME',
+  START_GAME: 'START_GAME',
+};
+
+export default actionTypes;
