+++ src/menu/index.js
@@ -3,5 +3,4 @@ import actionTypes from './actionTypes';
 import reducer from './reducer';
 import * as actions from './actions';
 
-
-export default { components, reducer, actions, actionTypes };
+export default { actions, actionTypes, components, reducer };
